Правки багов (#5)
* Правка бага с нескрывающимся меню при подключении к игре

* Спавн игроков в лобби без стандартного спавнера

* Правка бага с направлением движения пня при поворотах

+++ assets/scripts/characters/Character.cs
@@ -93,7 +93,7 @@ public partial class Character : CharacterBody3D
 
 	private float GetRotatedSideDirection(Vector3 direction)
 	{
-		return RotationDegrees.Y switch
+		return (int)RotationDegrees.Y switch
 		{
 			>= 0 and < 90 => direction.X,
 			>= 90 and < 180 => -direction.Z,
